requirements.txt
streamlit
pandas
plotly
streamlit-plotly-events
altair
openpyxl
openai
python-dotenv
tabulate